test: consolidate L1/L2 regularizer tests into test_lp_norms.py
Merge test_l1_norm.py and test_l2_norm.py into a single test_lp_norms.py
built on a shared BaseLpNormTest suite (mirroring base_linear_model_test.py),
with per-norm subclasses supplying the atom class, expected norm values, and
their own prox cases. Package-level imports keep it decoupled from the
upcoming source-file consolidation.
